refactor: orderStatus move from entities to valueObjects #5
